backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9
psycopg[binary]==3.1.18
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
pydantic[email]==2.5.0
pydantic-settings==2.1.0
redis==5.0.1
celery==5.3.4
httpx==0.25.2
aiohttp==3.9.1
python-dotenv==1.0.0
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
flake8==6.1.0
mypy==1.7.1
google-generativeai==0.8.1
feedparser==6.0.10
beautifulsoup4==4.12.2
lxml==4.9.3
docker==6.1.3
pygments==2.17.2

